chage menucontroller file to displayModule and create a displayCtrl module
