+++ index.html
@@ -92,6 +92,7 @@
     </main>
 </body>
 
-<script src="./js/menuroute.js" type="module"></script>
+<!--<script src="js/displayController.js" type="module"></script>-->
+<script src="js/displayModule.js" type="module"></script>
 
 </html>
